fixed warnings, added cmake options

+++ seedimg-filters-core/seedimg-filters-core.cpp
@@ -10,9 +10,10 @@ void seedimg::filters::grayscale(std::unique_ptr<seedimg::img> &inp_img,
   if (lightness) {
     for (auto &row : inp_img->get_data()) {
       for (auto &pix : row) {
-        uint8_t linear = (0.2126 * (pix.r / 255.0) + 0.7152 * (pix.g / 255.0) +
-                          0.0722 * (pix.b / 255.0)) *
-                         255;
+        uint8_t linear = static_cast<uint8_t>((0.2126 * (pix.r / 255.0) +
+                                               0.7152 * (pix.g / 255.0) +
+                                               0.0722 * (pix.b / 255.0)) *
+                                              255);
         pix = {linear, linear, linear, pix.a};
       }
     }
